Refactor to leave some room for the writer
